pyproject.toml
[tool.poetry]
name = "cch-python-slack"
version = "0.1.0"
description = ""
authors = ["cch-lion <ckdgus11021@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "cch_python_slack"}]

[tool.poetry.dependencies]
python = "^3.11"
slack-sdk = "^3.21.3"
black = "^23.7.0"
python-dotenv = "^1.0.0"
httpx = "^0.24.1"
selectolax = "^0.3.15"
playwright = "^1.36.0"
ptpython = "^3.0.23"
pytest-playwright = "^0.3.3"
poetry-dotenv-plugin = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
